Fix integer literal ambiguity with explicit static_cast to int64_t

+++ tests/test_hybrid_map.cpp
@@ -68,7 +68,7 @@ TEST(MemoryAndLifecycle, ClearForcesImmediateHeapReleaseInMapMode) {
     CHECK_EQ(map.size(), static_cast<size_t>(0));
 
     // We check that the mode has reset to vector and is ready to accept data again
-    map.add("dev", "new_param", ParameterValue(42LL));
+    map.add("dev", "new_param", ParameterValue(static_cast<int64_t>(42)));
     CHECK_EQ(map.size(), static_cast<size_t>(1));
 }
